Update Chp1-Ex2.cpp
A problem arises because when a user enters a name longer than 20 characters which results in leftover characters in the input stream. Since there are leftover characters in the buffer, cin cannot parse a valid float, and the operation fails and the value of gpa remains in its initialized value of 0.0

The setw(20) manipulator in C++ limits how many characters are extracted from the input stream and stored in the provided buffer (in this case, name). However, it does not consume or remove any characters beyond this limit from the input buffer. 

To fix this issue, we can clear the input stream after reading the name to ensure there are no leftover characters.